Revert "Revert "Merge branch 'master' of github.com:anl-polaris/polaris-hpc""
This reverts commit 1d1f99a66a85864ea84eb8b8978a57d41651e243.

+++ convergence/run_convergence.py
@@ -14,6 +14,7 @@ import csv
 import traceback
 import regression
 import modify_scenario
+import CSV_Utillities
 
 
 # +++++++++++++++++++++++++++++++++++++++
@@ -21,19 +22,35 @@ import modify_scenario
 # +++++++++++++++++++++++++++++++++++++++
 
 
-def run_polaris_local(results_dir, exe_name, scenario_file, num_threads):
+def run_polaris_local(results_dir, exe_name, scenario_file, num_threads, tail_app):
     # subprocess.call([exeName, arguments])
     out_file = open(str(results_dir / 'simulation_out.log'), 'w+')
     err_file = open(str(results_dir / 'simulation_err.log'), 'w+')
     proc = subprocess.Popen([str(exe_name), str(scenario_file), num_threads], stdout=out_file, stderr=subprocess.PIPE)
+    #tail_proc = run_tail_application( tail_app, results_dir)
     for line in proc.stderr:
         sys.stdout.write(str(line))
         err_file.write(str(line))
     proc.wait()
     out_file.close()
     err_file.close()
+    #tail_proc.terminate()
     if proc.returncode != 0:
         sys.exit("POLARIS did not execute correctly")
+
+
+def run_tail_application(tail_app, results_dir):
+    file_to_tail = results_dir / 'simulation_out.log'
+    if 'linux' in sys.platform:
+        print('Running linux tail app: %s %s' % (tail_app, file_to_tail))
+        tail_proc = subprocess.Popen([tail_app, str(file_to_tail)], shell=True)
+    elif 'windows' in sys.platform:
+        print('Running windows tail app: %s %s' % (tail_app, file_to_tail))
+        tail_proc = subprocess.Popen([tail_app, str(file_to_tail)])
+    else:
+        print("Unable to start tail application: %s %s" % (tail_app, file_to_tail) )
+
+    return tail_proc
 
 
 def copyreplacefile(filename, dest_dir):
@@ -45,10 +62,10 @@ def copyreplacefile(filename, dest_dir):
 
 def execute_sql_script(db_name, script):
     print('Executing Sqlite3 script: %s on database: %s' % (db_name, script))
-    with open(script, 'r') as sql_file:
+    with open(str(script), 'r') as sql_file:
         sql_script = sql_file.read()
 
-    db = sqlite3.connect(db_name)
+    db = sqlite3.connect(str(db_name))
     cursor = db.cursor()
     try:
         cursor.executescript(sql_script)
@@ -63,11 +80,11 @@ def execute_sql_script(db_name, script):
 
 
 def dump_table_to_csv(db, table, csv_name, loop):
-    db_input = sqlite3.connect(db)
+    db_input = sqlite3.connect(str(db))
     sql3_cursor = db_input.cursor()
     query = 'SELECT * FROM ' + table
     sql3_cursor.execute(query)
-    with open(csv_name, 'w') as out_csv_file:
+    with open(str(csv_name), 'w') as out_csv_file:
         csv_out = csv.writer(out_csv_file, lineterminator='\n')  # gets rid of blank lines - defaults to \r\n
         if loop == 0:
             # write header
@@ -85,80 +102,60 @@ def append_file(src, tar):
         tar_file.close()
 
 
-def run_conv(control_file_name):
+def run_conv(control_file_name, data_directory):
     control_file = Path(control_file_name)
     if not control_file.exists():
-        print('ERROR: %s does not exist!' % control_file.name)
-        return None  # or: raise
+        print(f'ERROR: \'{control_file.name}\' does not exist!')
+        sys.exit(-1)
 
     with control_file.open() as f:
         try:
             json_data = json.load(f)
-        except ValueError as e:
-            print('invalid json: %s' % e)
-            return None  # or: raise
+        except ValueError as err:
+            print(f'invalid json: {err}')
+            sys.exit(-1)
 
     json_formatted_str = json.dumps(json_data, indent=2)
 
     print(json_formatted_str)
 
-    # exe_name = sys.argv[1]
-    # results_dir = sys.argv[2]
-
-    # -- SET THE MAIN AND DTA RUN SCENARIO FILES
-    # set python_path=C:\Program Files\Python36\
-    # set sqlite3_path=%CD%\third_party\Sqlite3\
     # set tail_app=%CD%\third_party\baretail.exe
     # set cloud_backup_path=\\vms-fs2\VMS_FY19_SMART_Runs\Chicago_ABM_Convergence2018\Chicago_202000406_results\
     #
-    # set SCENARIO_MAIN_INIT=scenario_abm_init_10per.json
-    # set SCENARIO_MAIN=scenario_abm.json
 
-    # python_path = json_data["python_path"]
-    # sqlite3_path = json_data["sqlite3_path"]
-    # tail_app = json_data["tail_app"]
-    # cloud_backup_path = json_data["cloud_backup_path"]
+    tail_app = json_data["tail_app"]
+    database_base_name = json_data["database_base_name"]
 
-    database_name = json_data["database_name"]
+    # -- SET THE POLARIS RUN INFORMATION
+    exe_name = Path(json_data["model"])
+    #data_dir = Path(json_data["data"])
+    data_dir = Path(data_directory)
+    if not data_dir.exists():
+        print(f'ERROR: \'{data_directory}\' does not exist!')
+        sys.exit(-1)
+    results_dir = data_dir / json_data["results_dir"]
+    num_threads = json_data["num_threads"]
+    scripts_dir = Path(json_data["scripts_dir"])
 
+    # -- SET THE MAIN AND DTA RUN SCENARIO FILES
     scenario_main_init = json_data["scenario_main_init"]
     scenario_main = json_data["scenario_main"]
 
-    exe_name = Path(json_data["model"])
-    data_dir = Path(json_data["data"])
-    results_dir = data_dir / json_data["results_dir"]
-    # scenario_file = json_data["model_scenario"]
-    num_threads = json_data["num_threads"]
-    # standard_dir = json_data["standard_dir"]
-
     # -- ENTER THE NUMBER OF MAIN RUNS AND DTA RUNS PER MAIN RUN
-
     num_abm_runs = json_data["num_abm_runs"]
 
     # -- SET THE OUTPUT DIRECTORIES AS SPECIFIED IN THE SCENARIO FILES
-
     # output_directories = json_data["output_directories"]
-
-    # set OUT_DIR_MAIN=linux_vs17_chicago2018_abm
-    #
-    # -- SET THE POLARIS RUN INFORMATION
-    # set THREADS=20
-    # set DB=chicago2018
-    # set POLARIS_EXE=bin_new\Integrated_Model.exe
+    cloud_backup_path = Path(json_data["cloud_backup_path"])
 
     # -------------------------------------------------------------------------------------------
     # Do not modify below here
     # -------------------------------------------------------------------------------------------
 
-    # cwd = os.getcwd()
-    # app_dir = Path.cwd()
     print(data_dir)
     os.chdir(str(data_dir))
     working_dir = Path.cwd()
 
-    # del artificial_count.csv
-    # del gap_calculations.csv
-    # del gap_breakdown.csv
     if Path("artificial_count.csv").exists():
         os.remove("artificial_count.csv")
     if Path("gap_calculations.csv").exists():
@@ -167,16 +164,19 @@ def run_conv(control_file_name):
         os.remove("gap_breakdown.csv")
 
     # store the original inputs
-    demand_db_name = database_name + "-Demand.sqlite"
-    result_db_name = database_name + "-Result.sqlite"
+    demand_db_name = database_base_name + "-Demand.sqlite"
+    result_db_name = database_base_name + "-Result.sqlite"
     demand_db = working_dir / "backup" / demand_db_name
-    skim_file_name = "highway_skim_file.bin"
-    skim_file = working_dir / "backup" / skim_file_name
+    highway_skim_file_name = "highway_skim_file.bin"
+    highway_skim_file = working_dir / "backup" / highway_skim_file_name
+    transit_skim_file_name = "transit_skim_file.bin"
+    transit_skim_file = working_dir / "backup" / transit_skim_file_name
     output_file = results_dir / "simulation_out.log"
 
     copyreplacefile(demand_db, working_dir)
-    copyreplacefile(skim_file, working_dir)
-    print("Polaris output goes here: \'" + str(output_file) + "\'")
+    copyreplacefile(highway_skim_file, working_dir)
+    copyreplacefile(transit_skim_file, working_dir)
+    print(f"Polaris output goes here: {str(output_file)}")
 
     # Process main ABM run
 
@@ -184,24 +184,26 @@ def run_conv(control_file_name):
     result_paths = []
 
     for loop in range(0, int(num_abm_runs)):
+        print("\n------------------------------------------------------------------------")
         # Create results Directory if don't exist
         if not results_dir.exists():
             os.mkdir(str(results_dir))
-            print("Directory ", results_dir, " Created ")
+            print(f"Directory:  {results_dir} Created ")
         else:
-            print("Directory ", results_dir, " already exists")
+            print(f"Directory: {results_dir} already exists")
 
         if loop == 0:
             scenario_file = scenario_main_init
+            print(f"Running Polaris SCENARIO_MAIN_INIT instance {loop} - see {results_dir / 'simulation_out.log'}")
         else:
             scenario_file = scenario_main
             modify_scenario.modify(scenario_main, "time_dependent_routing_weight_factor", 1.0)
-            print("\nRunning Polaris SCENARIO_MAIN instance ", loop, " - see", results_dir / 'simulation_out.log')
+            print(f"Running Polaris SCENARIO_MAIN instance #{loop} - see {results_dir / 'simulation_out.log'}")
 
         arguments = scenario_file + ' ' + num_threads
-        print('Executing \'' + str(exe_name) + ' ' + arguments + '\'')
+        print(f'Executing \'{str(exe_name)} {arguments}\'')
 
-        run_polaris_local(results_dir, exe_name, scenario_file, num_threads)
+        run_polaris_local(results_dir, exe_name, scenario_file, num_threads, tail_app)
 
         all_subdirs = [d for d in os.listdir('.') if os.path.isdir(d)]
         latest_subdir = Path(max(all_subdirs, key=os.path.getmtime))
@@ -211,7 +213,7 @@ def run_conv(control_file_name):
 
         # move the output files (now that we know where the simulation files were created)
         results_dir_moved = working_dir / latest_subdir / json_data["results_dir"]
-        print('Moving: ', results_dir, ' to:', results_dir_moved)
+        print(f'Moving: {results_dir} to: {results_dir_moved}')
         shutil.move(str(results_dir), str(results_dir_moved))
         # os.rename('./simulation_out.log', simulated_dir + '/simulation_out.log')
         # os.rename('./simulation_err.log', simulated_dir + '/simulation_err.log')
@@ -219,15 +221,17 @@ def run_conv(control_file_name):
         # copy local results back to the main run directory for the next run
         copyreplacefile(working_dir / latest_subdir / demand_db_name, working_dir)
         copyreplacefile(working_dir / latest_subdir / result_db_name, working_dir)
-        copyreplacefile(working_dir / latest_subdir / skim_file_name, working_dir)
+        if loop > 0:
+            copyreplacefile(working_dir / latest_subdir / highway_skim_file_name, working_dir)
+            #copyreplacefile(working_dir / latest_subdir / transit_skim_file_name, working_dir)
 
         # %sqlite3_path%sqlite3.exe "%WORKDIR%\%DB%-Demand.sqlite" < clean_db_after_abm_for_abm.sql
-        execute_sql_script(working_dir / demand_db_name, working_dir / "clean_db_after_abm_for_abm.sql")
+        execute_sql_script(working_dir / demand_db_name, scripts_dir / "clean_db_after_abm_for_abm.sql")
 
         # ren %WORKDIR%\!out_local!\summary.csv summary_abm_%%S.csv
         latest_demand_db = working_dir / latest_subdir / demand_db_name
-        execute_sql_script(latest_demand_db, working_dir / "wtf_baseline_analysis_25Per_calibrate.sql")
-        execute_sql_script(latest_demand_db, working_dir / "gap_calculate.sql")
+        execute_sql_script(latest_demand_db, scripts_dir / "wtf_baseline_analysis_25Per_calibrate.sql")
+        execute_sql_script(latest_demand_db, scripts_dir / "gap_calculate.sql")
         # execute_sql_script(working_dir / latest_subdir / demand_db_name, working_dir / "output_to_csv.sql")
         dump_table_to_csv(latest_demand_db, "artificial_count", working_dir / "artificial_count_temp.csv", loop)
         dump_table_to_csv(latest_demand_db, "gap_calculations", working_dir / "gap_calculations_temp.csv", loop)
@@ -243,22 +247,37 @@ def run_conv(control_file_name):
         os.remove(working_dir / "gap_breakdown_temp.csv")
 
         if loop > 0:
-            print('Checking convergence on \'' + str(latest_subdir) + '\'')
+            print(f'Checking convergence on {str(latest_subdir)}')
             regression.regression(result_paths[loop - 1].name, result_paths[loop].name)
 
-        # make sure results_dir exists
-        # if not os.path.exists(results_dir):
-        #    os.makedirs(results_dir)
+        # merge in_network data into a composite summary file
+        CSV_Utillities.append_column(working_dir / latest_subdir / 'summary.csv', working_dir / 'in_network.csv', loop, 4, str(latest_subdir))
 
-        # copyfile('./Regression_Report.html', results_dir + '/Regression_Report.html')
-        # copyfile(simulated_dir + '/in_network.png', results_dir + '/in_network.png')
-        # copyfile(simulated_dir + '/simulation_out.log', results_dir + '/simulation_out.log')
-        # copyfile(simulated_dir + '/simulation_err.log', results_dir + '/simulation_err.log')
+        # 	xcopy "artificial_count.csv" "%cloud_backup_path%" /y /i
+        # 	xcopy "gap_calculations.csv" "%cloud_backup_path%" /y /i
+        # 	xcopy "gap_breakdown.csv" "%cloud_backup_path%" /y /i
+        # 	xcopy "%WORKDIR%\!out_local!\summary_abm_%%S.csv" "%cloud_backup_path%" /y /i
+        # 	xcopy "%WORKDIR%\!out_local!\%DB%-Demand.sqlite" "%cloud_backup_path%\!out_local!\" /y /i
+        # 	xcopy "%WORKDIR%\!out_local!\%DB%-Result.sqlite" "%cloud_backup_path%\!out_local!\" /y /i
+        # 	xcopy "%WORKDIR%\!out_local!\highway_skim_file.bin" "%cloud_backup_path%\!out_local!\" /y /i
+        if not os.path.exists(cloud_backup_path / latest_subdir):
+            os.makedirs(cloud_backup_path / latest_subdir)
+        shutil.copyfile(str(working_dir / 'artificial_count.csv'), str(cloud_backup_path / 'artificial_count.csv'))
+        shutil.copyfile(str(working_dir / 'gap_calculations.csv'), str(cloud_backup_path / 'gap_calculations.csv'))
+        shutil.copyfile(str(working_dir / 'gap_breakdown.csv'), str(cloud_backup_path / 'gap_breakdown.csv'))
+        shutil.copyfile(str(working_dir / 'in_network.csv'), str(cloud_backup_path / 'in_network.csv'))
+        shutil.copyfile(str(working_dir / latest_subdir / 'summary.csv'), str(cloud_backup_path / latest_subdir / 'summary.csv'))
+        shutil.copyfile(str(working_dir / latest_subdir / demand_db_name), str(cloud_backup_path / latest_subdir / demand_db_name))
+        shutil.copyfile(str(working_dir / latest_subdir / result_db_name), str(cloud_backup_path / latest_subdir / result_db_name))
+        if loop > 0:
+            shutil.copyfile(str(working_dir / latest_subdir / 'highway_skim_file.bin'), str(cloud_backup_path / latest_subdir / 'highway_skim_file.bin'))
+            #shutil.copyfile(str(working_dir / latest_subdir / 'transit_skim_file.bin'), str(cloud_backup_path / latest_subdir / 'transit_skim_file.bin'))
+            shutil.copyfile(str(working_dir / latest_subdir / 'in_network.png'), str(cloud_backup_path / latest_subdir / 'in_network.png'))
 
 
 if __name__ == '__main__':
-    if len(sys.argv) < 2:
-        print('Usage %s <json_control_file>' % (sys.argv[0]))
+    if len(sys.argv) < 3:
+        print(f'Usage {sys.argv[0]} <json_control_file> <data_directory>')
         sys.exit(-1)
 
-    run_conv(sys.argv[1])
+    run_conv(sys.argv[1], sys.argv[2])
